Create Safari Books Online Search Bookmarklet
Packt Publishing books tend to be lower quality, but they try
to make up with it by volume.  As a user of Safari Books
Online, I want to eliminate Packt Publishing books from
my search in order to remove alot of noise.  This
is my attempt to create that ability.