Add Eclipse files to gitignore
I disagree that .project and .classpath should be committed to source control.  This does nothing but cause issues.  I prefer IDE settings to be the domain of the developer (and are subject to versions and personal preferences).  Committing these files causes nothing but headaches, and I think it was a mistake for the gitignore project to remove theseAdd Eclipse files to gitignore

I disagree that .project and .classpath should be committed to source control.  This does nothing but cause issues.  I prefer IDE settings to be the domain of the developer (and are subject to IDE versions and personal preferences).  Committing these files causes nothing but headaches, and I think it was a mistake for the gitignore project to remove these.
.